Step 4 (#340)
* Lotto 클래스 리팩토링 (LottoResults 빼냄)
생명주기가 다르다.

* 예외처리와 LottoInputView 추가 (사용자 입력 처리)

+++ src/main/java/dto/LottoResultDto.java
@@ -5,9 +5,9 @@ import domain.LottoResult;
 public class LottoResultDto implements Comparable<LottoResultDto> {
 
     private final LottoResult lottoResult;
-    private final int count;
+    private final long count;
 
-    public LottoResultDto(LottoResult lottoResult, int count) {
+    public LottoResultDto(LottoResult lottoResult, long count) {
         this.lottoResult = lottoResult;
         this.count = count;
     }
@@ -16,7 +16,7 @@ public class LottoResultDto implements Comparable<LottoResultDto> {
         return lottoResult;
     }
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
